requirements.txt
requests==2.31.0
slack_sdk==3.27.1
python-dotenv==1.0.1
openai==1.14.0
httpx<0.28.0
schedule==1.2.1
